refactor: response 파일 이름 개발 컨벤션에 맞게 변경
